Replace forced database drop in `pytest` teardown with a safer version
- Originally ran `DROP DATABASE testDbName WITH (FORCE)` SQL statement since
the Flask app was still technically connected to the database even during teardown
and appending `WITH (FORCE)` would cause all connections to be dropped, enabling
the database to be deleted.
- HOWEVER, since the app's DB connection is available from its `__init__.py`,
the `db` connection can be imported and used as `sqlalchemy`, gaining access to
the connection pool to manually and more safely dispose of the Flask app's database
session to cleanly drop the database.
  - New SQL statement can now more simply be `DROP DATABASE testDbName`